chore(database): setup database - setup database for all endpoints - integrate the already created endpoints [Delivers #163673507]
