+++ .env
@@ -1,0 +1,5 @@
+PGUSER='postgres'
+PGHOST='localhost'
+PGDATABASE='template1'
+PGPASSWORD='Spectre007'
+PGPORT=5432
