+++ middlewares/officeMiddleware.js
@@ -1,8 +1,13 @@
 class OfficeMiddleware {
-  static createOfficeCheck(req) {
+  static createOfficeCheck(req, res, next) {
     if (!req.body.type || !req.body.name) {
-      req.error = true;
+      return res.status(404).json({
+        status: 404,
+        error: 'Provide type and name of the office.',
+      });
     }
+
+    next();
   }
 }
