+++ middlewares/partyMiddleware.js
@@ -1,14 +1,24 @@
 class PartyMiddleware {
-  static createPartyCheck(req) {
+  static createPartyCheck(req, res, next) {
     if (!req.body.name || !req.body.hqAddress || !req.body.logoUrl) {
-      req.error = true;
+      return res.status(404).json({
+        status: 404,
+        error: 'Provide name, address and logo of the party',
+      });
     }
+
+    next();
   }
 
-  static editSpecificPartyCheck(req) {
+  static editSpecificPartyCheck(req, res, next) {
     if (!req.body.name) {
-      req.error = true;
+      return res.status(404).json({
+        status: 404,
+        error: 'Provide new name of the party',
+      });
     }
+
+    next();
   }
 }
