+++ model/db/db.js
@@ -1,0 +1,8 @@
+import { Pool } from 'pg';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const db = new Pool();
+
+export default db;
