+++ model/officeQueries/getOffice.js
@@ -1,0 +1,33 @@
+import db from '../db/db';
+
+const getOffice = (req, res) => {
+  const id = parseInt(req.params.id, 10);
+
+  const text = 'SELECT * FROM office WHERE id=$1';
+  const values = [id];
+
+  db.query(text, values, (error, results) => {
+    if (error) {
+      return res.status(500).json({
+        status: 500,
+        error: error.detail,
+      });
+    }
+
+    if (results.rowCount > 0) {
+      res.status(200).json({
+        status: 200,
+        data: [
+          results.rows[0],
+        ],
+      });
+    } else {
+      res.status(404).json({
+        status: 404,
+        message: 'No office matches the id',
+      });
+    }
+  });
+};
+
+export default getOffice;
