--- models/datastore/officeData.js
@@ -1,3 +1,0 @@
-const offices = [];
-
-export default offices;
