--- models/datastore/partyData.js
@@ -1,3 +1,0 @@
-const parties = [];
-
-export default parties;
