+++ package.json
@@ -44,11 +44,15 @@
     "mocha": "^5.2.0",
     "nodemon": "^1.18.9",
     "nyc": "^13.1.0",
-    "supertest": "^3.4.2"
+    "supertest": "^3.4.2",
+    "trim-request": "^1.0.6"
   },
   "dependencies": {
     "body-parser": "^1.18.3",
     "express": "^4.16.4",
-    "morgan": "^1.9.1"
+    "pg": "^7.8.0",
+    "dotenv": "^6.2.0",
+    "morgan": "^1.9.1",
+    "trim-request": "^1.0.6"
   }
 }
